rename class py file. revise fit()
