+++ feature_engine/creation/__init__.py
@@ -5,7 +5,7 @@ variables in the dataframe.
 from .cyclical_features import CyclicalFeatures
 from .math_features import MathFeatures
 from .relative_features import RelativeFeatures
-from .decision_tree import DecisionTreeFeatures
+from .decision_tree_features import DecisionTreeFeatures
 
 # FIXME: remove in version 1.4
 from .combine_with_reference_feature import CombineWithReferenceFeature
